move incremental fetching tests to their own class
