+++ tests/Keboola/DbExtractor/MySQLTest.php
@@ -19,7 +19,7 @@ class MySQLTest extends AbstractMySQLTest
         $config['action'] = 'testConnection';
         unset($config['parameters']['tables']);
         $config['parameters']['db']['networkCompression'] = true;
-        
+
         $app = $this->createApplication($config);
         $result = $app->run();
 
@@ -206,7 +206,7 @@ class MySQLTest extends AbstractMySQLTest
 
         $this->assertEquals('success', $result['status']);
         $this->assertCount(3, $result['tables']);
-        
+
         $expectedData = array (
             0 =>
                 array (
@@ -1175,235 +1175,6 @@ class MySQLTest extends AbstractMySQLTest
         $this->assertEquals(['weird_I_d'], $manifest['primary_key']);
     }
 
-    public function testIncrementalFetchingByTimestamp(): void
-    {
-        $config = $this->getIncrementalFetchingConfig();
-        $config['parameters']['incrementalFetchingColumn'] = 'timestamp';
-        $this->createAutoIncrementAndTimestampTable();
-
-        $result = ($this->createApplication($config))->run();
-
-        $this->assertEquals('success', $result['status']);
-        $this->assertEquals(
-            [
-                'outputTable' => 'in.c-main.auto-increment-timestamp',
-                'rows' => 2,
-            ],
-            $result['imported']
-        );
-
-        //check that output state contains expected information
-        $this->assertArrayHasKey('state', $result);
-        $this->assertArrayHasKey('lastFetchedRow', $result['state']);
-        $this->assertNotEmpty($result['state']['lastFetchedRow']);
-
-        sleep(2);
-        // the next fetch should be empty
-        $emptyResult = ($this->createApplication($config, $result['state']))->run();
-        $this->assertEquals(0, $emptyResult['imported']['rows']);
-
-        sleep(2);
-        //now add a couple rows and run it again.
-        $this->pdo->exec('INSERT INTO auto_increment_timestamp (`weird-Name`) VALUES (\'charles\'), (\'william\')');
-
-        $newResult = ($this->createApplication($config, $result['state']))->run();
-
-        //check that output state contains expected information
-        $this->assertArrayHasKey('state', $newResult);
-        $this->assertArrayHasKey('lastFetchedRow', $newResult['state']);
-        $this->assertGreaterThan(
-            $result['state']['lastFetchedRow'],
-            $newResult['state']['lastFetchedRow']
-        );
-        $this->assertEquals(2, $newResult['imported']['rows']);
-    }
-
-    public function testIncrementalFetchingByDatetime(): void
-    {
-        $config = $this->getIncrementalFetchingConfig();
-        $config['parameters']['incrementalFetchingColumn'] = 'datetime';
-        $config['parameters']['table']['tableName'] = 'auto_increment_timestamp_withFK';
-        $config['parameters']['outputTable'] = 'in.c-main.auto-increment-timestamp-with-fk';
-        $this->createAutoIncrementAndTimestampTable();
-        $this->createAutoIncrementAndTimestampTableWithFK();
-
-        $result = ($this->createApplication($config))->run();
-
-        $this->assertEquals('success', $result['status']);
-        $this->assertEquals(
-            [
-                'outputTable' => 'in.c-main.auto-increment-timestamp-with-fk',
-                'rows' => 1,
-            ],
-            $result['imported']
-        );
-
-        //check that output state contains expected information
-        $this->assertArrayHasKey('state', $result);
-        $this->assertArrayHasKey('lastFetchedRow', $result['state']);
-        $this->assertNotEmpty($result['state']['lastFetchedRow']);
-
-        sleep(2);
-        // the next fetch should be empty
-        $emptyResult = ($this->createApplication($config, $result['state']))->run();
-        $this->assertEquals(0, $emptyResult['imported']['rows']);
-
-        sleep(2);
-        //now add a couple rows and run it again.
-        $this->pdo->exec('INSERT INTO auto_increment_timestamp_withFK (`random_name`) VALUES (\'charles\'), (\'william\')');
-
-        $newResult = ($this->createApplication($config, $result['state']))->run();
-
-        //check that output state contains expected information
-        $this->assertArrayHasKey('state', $newResult);
-        $this->assertArrayHasKey('lastFetchedRow', $newResult['state']);
-        $this->assertGreaterThan(
-            $result['state']['lastFetchedRow'],
-            $newResult['state']['lastFetchedRow']
-        );
-        $this->assertEquals(2, $newResult['imported']['rows']);
-    }
-
-    public function testIncrementalFetchingByAutoIncrement(): void
-    {
-        $config = $this->getIncrementalFetchingConfig();
-        $config['parameters']['incrementalFetchingColumn'] = '_weird-I-d';
-        $this->createAutoIncrementAndTimestampTable();
-
-        $result = ($this->createApplication($config))->run();
-
-        $this->assertEquals('success', $result['status']);
-        $this->assertEquals(
-            [
-                'outputTable' => 'in.c-main.auto-increment-timestamp',
-                'rows' => 2,
-            ],
-            $result['imported']
-        );
-
-        //check that output state contains expected information
-        $this->assertArrayHasKey('state', $result);
-        $this->assertArrayHasKey('lastFetchedRow', $result['state']);
-        $this->assertEquals(2, $result['state']['lastFetchedRow']);
-
-        sleep(2);
-        // the next fetch should be empty
-        $emptyResult = ($this->createApplication($config, $result['state']))->run();
-        $this->assertEquals(0, $emptyResult['imported']['rows']);
-
-        sleep(2);
-        //now add a couple rows and run it again.
-        $this->pdo->exec('INSERT INTO auto_increment_timestamp (`weird-Name`) VALUES (\'charles\'), (\'william\')');
-
-        $newResult = ($this->createApplication($config, $result['state']))->run();
-
-        //check that output state contains expected information
-        $this->assertArrayHasKey('state', $newResult);
-        $this->assertArrayHasKey('lastFetchedRow', $newResult['state']);
-        $this->assertEquals(4, $newResult['state']['lastFetchedRow']);
-        $this->assertEquals(2, $newResult['imported']['rows']);
-    }
-
-    public function testIncrementalFetchingLimit(): void
-    {
-        $config = $this->getIncrementalFetchingConfig();
-        $config['parameters']['incrementalFetchingLimit'] = 1;
-        $this->createAutoIncrementAndTimestampTable();
-
-        $result = ($this->createApplication($config))->run();
-
-        $this->assertEquals('success', $result['status']);
-        $this->assertEquals(
-            [
-                'outputTable' => 'in.c-main.auto-increment-timestamp',
-                'rows' => 1,
-            ],
-            $result['imported']
-        );
-
-        //check that output state contains expected information
-        $this->assertArrayHasKey('state', $result);
-        $this->assertArrayHasKey('lastFetchedRow', $result['state']);
-        $this->assertEquals(1, $result['state']['lastFetchedRow']);
-
-        sleep(2);
-        // the next fetch should contain the second row
-        $result = ($this->createApplication($config, $result['state']))->run();
-        $this->assertEquals(
-            [
-                'outputTable' => 'in.c-main.auto-increment-timestamp',
-                'rows' => 1,
-            ],
-            $result['imported']
-        );
-
-        //check that output state contains expected information
-        $this->assertArrayHasKey('state', $result);
-        $this->assertArrayHasKey('lastFetchedRow', $result['state']);
-        $this->assertEquals(2, $result['state']['lastFetchedRow']);
-    }
-
-    public function testIncrementalOrdering(): void
-    {
-        $this->createAutoIncrementAndTimestampTable();
-        $config = $this->getIncrementalFetchingConfig();
-
-        $result = ($this->createApplication($config))->run();
-        $outputCsvFile = iterator_to_array(
-            new CsvFile(
-                $this->dataDir . '/out/tables/' . $result['imported']['outputTable'] . '.csv'
-            )
-        );
-
-        $previousId = 0;
-        foreach ($outputCsvFile as $key => $row) {
-            $this->assertGreaterThan($previousId, (int) $row[0]);
-            $previousId = (int) $row[0];
-        }
-    }
-
-    /**
-     * @dataProvider invalidColumnProvider
-     */
-    public function testIncrementalFetchingInvalidColumns(string $column, string $expectedExceptionMessage): void
-    {
-        $this->createAutoIncrementAndTimestampTable();
-        $config = $this->getIncrementalFetchingConfig();
-        $config['parameters']['incrementalFetchingColumn'] = $column;
-
-        $this->setExpectedException(UserException::class, $expectedExceptionMessage);
-        ($this->createApplication($config))->run();
-    }
-
-    public function invalidColumnProvider(): array
-    {
-        return [
-            'column does not exist' => [
-                "fakeCol",
-                "Column [fakeCol] specified for incremental fetching was not found in the table",
-            ],
-            'column exists but is not auto-increment nor updating timestamp so should fail' => [
-                "weird-Name",
-                "Column [weird-Name] specified for incremental fetching is not an auto increment column or a timestamp",
-            ],
-        ];
-    }
-
-    public function testIncrementalFetchingInvalidConfig(): void
-    {
-        $this->createAutoIncrementAndTimestampTable();
-        $config = $this->getIncrementalFetchingConfig();
-        $config['parameters']['query'] = 'SELECT * FROM auto_increment_timestamp';
-        unset($config['parameters']['table']);
-
-        try {
-            $result = ($this->createApplication($config))->run();
-            $this->fail('cannot use incremental fetching with advanced query, should fail.');
-        } catch (UserException $e) {
-            $this->assertStringStartsWith("Invalid Configuration", $e->getMessage());
-        }
-    }
-
     public function testRunWithNetworkCompression(): void
     {
         $config = $this->getIncrementalFetchingConfig();
@@ -1421,22 +1192,6 @@ class MySQLTest extends AbstractMySQLTest
         $this->assertArrayHasKey('state', $result);
         $this->assertArrayHasKey('lastFetchedRow', $result['state']);
         $this->assertEquals(2, $result['state']['lastFetchedRow']);
-    }
-
-    private function getIncrementalFetchingConfig(): array
-    {
-        $config = $this->getConfigRow(self::DRIVER);
-        unset($config['parameters']['query']);
-        $config['parameters']['table'] = [
-            'tableName' => 'auto_increment_timestamp',
-            'schema' => 'test',
-        ];
-        $config['parameters']['incremental'] = true;
-        $config['parameters']['name'] = 'auto-increment-timestamp';
-        $config['parameters']['outputTable'] = 'in.c-main.auto-increment-timestamp';
-        $config['parameters']['primaryKey'] = ['_weird-I-d'];
-        $config['parameters']['incrementalFetchingColumn'] = '_weird-I-d';
-        return $config;
     }
 
     public function testDBSchemaMismatchConfigRowWithNoName(): void
